Add grobid config as command line argument

+++ Mamorx/MAMORX/schemas.py
@@ -12,7 +12,7 @@ class PDFReviewResult(TypedDict):
 
 class ReviewResult(TypedDict):
     review_content: str
-    time_elapsed: int
+    time_elapsed: float
 
 
 class PaperReviewResult(TypedDict):
